+++ examples/using_geodata.py
@@ -14,16 +14,18 @@ for block_ref in msp.query("INSERT[layer=='VO_body+cisla']"):
     pprint(block_ref.dxfattribs())
     # GEODATA handle is stored in the associated BLOCK_RECORD
     block_record = doc.block_records.get(block_ref.dxf.name)
-    try:  # GEODATA handle is stored in an extension dictionary
+
+    # GEODATA handle is stored in an extension dictionary
+    if block_record.has_extension_dict:
         xdict = block_record.get_extension_dict()
-    except ezdxf.DXFValueError:
+    else:
         continue
     try:  # GEODATA handle is stored by key ACAD_GEOGRAPHICDATA
         geodata_handle = xdict['ACAD_GEOGRAPHICDATA']
     except ezdxf.DXFKeyError:
         pass
     else:  # but this seems never the case for a block reference
-        print('uses GEODATA(#{})'.format(geodata_handle))
+        print(f'uses GEODATA(#{geodata_handle})')
 
 # GEODATA entity is only referenced by the model space
 geodata = msp.get_geodata()