+++ tests/test_01_dxf_entities/test_104_extension_dict.py
@@ -17,8 +17,8 @@ def entity(doc):
 
 
 def test_new_extension_dict(doc, entity):
-    assert entity.has_extension_dict() is False
-    xdict = entity.get_extension_dict()
+    assert entity.has_extension_dict is False
+    xdict = entity.new_extension_dict(doc)
     assert xdict.dictionary.dxftype() == 'DICTIONARY'
     assert len(xdict.dictionary) == 0
     placeholder = xdict.add_placeholder('TEST')
@@ -27,8 +27,8 @@ def test_new_extension_dict(doc, entity):
     assert 'TEST' in xdict.dictionary
 
 
-def test_direct_interface(entity):
-    xdict = entity.get_extension_dict()
+def test_direct_interface(doc, entity):
+    xdict = entity.new_extension_dict(doc)
     placeholder = xdict.add_placeholder('TEST')
     assert 'TEST' in xdict
     placeholder2 = xdict['TEST']
@@ -37,24 +37,28 @@ def test_direct_interface(entity):
 
 
 def test_copy_entity(doc, entity):
-    xdict = entity.get_extension_dict()  # create a new extension dict if not exists
+    try:
+        xdict = entity.get_extension_dict()  # create a new extension dict if not exists
+    except AttributeError:
+        xdict = entity.new_extension_dict(doc)
+
     xdict.add_placeholder('Test')
 
     new_entity = entity.copy()
     # copying of extension dict is not supported
-    assert new_entity.has_extension_dict() is False
+    assert new_entity.has_extension_dict is False
 
 
 def test_line_new_extension_dict(doc):
     msp = doc.modelspace()
     entity = msp.add_line((0, 0), (10, 0))
-    assert entity.has_extension_dict() is False
-    xdict = entity.get_extension_dict()
+    assert entity.has_extension_dict is False
+    xdict = entity.new_extension_dict(doc)
     dxf_dict = xdict.dictionary
     assert dxf_dict.dxftype() == 'DICTIONARY'
     assert dxf_dict.dxf.owner == entity.dxf.handle
     assert entity.has_app_data('{ACAD_XDICTIONARY') is False, 'extension dictionary is a separated storage'
-    assert entity.has_extension_dict() is True
+    assert entity.has_extension_dict is True
 
     xdict2 = entity.get_extension_dict()
     dxf_dict2 = xdict2.dictionary
@@ -64,7 +68,7 @@ def test_line_new_extension_dict(doc):
 def test_del_entity_with_ext_dict(doc):
     msp = doc.modelspace()
     entity = msp.add_line((0, 0), (10, 0))
-    xdict = entity.get_extension_dict()
+    xdict = entity.new_extension_dict(doc)
 
     objects = doc.objects
     assert xdict.dictionary in objects